Get rid of circular dependencies between Stats.Imported and Stats.Imported.Base
