+++ lib/plausible/stats/imported/base.ex
@@ -2,9 +2,40 @@ defmodule Plausible.Stats.Imported.Base do
   @moduledoc """
   A module for building the base of an imported stats query
   """
-  alias Plausible.Stats.{Query, Imported, Filters}
 
   import Ecto.Query
+
+  alias Plausible.Imported
+  alias Plausible.Stats.Filters
+  alias Plausible.Stats.Query
+
+  @property_to_table_mappings %{
+    "visit:source" => "imported_sources",
+    "visit:referrer" => "imported_sources",
+    "visit:utm_source" => "imported_sources",
+    "visit:utm_medium" => "imported_sources",
+    "visit:utm_campaign" => "imported_sources",
+    "visit:utm_term" => "imported_sources",
+    "visit:utm_content" => "imported_sources",
+    "visit:entry_page" => "imported_entry_pages",
+    "visit:exit_page" => "imported_exit_pages",
+    "visit:country" => "imported_locations",
+    "visit:region" => "imported_locations",
+    "visit:city" => "imported_locations",
+    "visit:device" => "imported_devices",
+    "visit:browser" => "imported_browsers",
+    "visit:browser_version" => "imported_browsers",
+    "visit:os" => "imported_operating_systems",
+    "visit:os_version" => "imported_operating_systems",
+    "event:page" => "imported_pages",
+    "event:name" => "imported_custom_events",
+    "event:props:url" => "imported_custom_events",
+    "event:props:path" => "imported_custom_events",
+
+    # NOTE: these properties can be only filtered by
+    "visit:screen" => "imported_devices",
+    "event:hostname" => "imported_pages"
+  }
 
   @goals_with_url Imported.goals_with_url()
   @goals_with_path Imported.goals_with_path()
@@ -23,15 +54,17 @@ defmodule Plausible.Stats.Imported.Base do
     url: :link_url
   }
 
+  def property_to_table_mappings(), do: @property_to_table_mappings
+
   def query_imported(site, query) do
     query
-    |> Imported.transform_filters()
+    |> transform_filters()
     |> decide_table()
     |> query_imported(site, query)
   end
 
   def query_imported(table, site, query) do
-    query = Imported.transform_filters(query)
+    query = transform_filters(query)
     import_ids = site.complete_import_ids
     %{first: date_from, last: date_to} = query.date_range
 
@@ -45,11 +78,49 @@ defmodule Plausible.Stats.Imported.Base do
     |> apply_filter(query)
   end
 
-  def decide_table(%Query{filters: [], property: nil}), do: "imported_visitors"
-  def decide_table(%Query{filters: [], property: "event:props:url"}), do: nil
-  def decide_table(%Query{filters: [], property: "event:props:path"}), do: nil
+  def decide_table(query) do
+    query
+    |> transform_filters()
+    |> do_decide_table()
+  end
 
-  def decide_table(%Query{filters: filters, property: "event:props:url"}) do
+  defp transform_filters(query) do
+    new_filters =
+      query.filters
+      |> Enum.reject(fn
+        [:is, "event:name", "pageview"] -> true
+        _ -> false
+      end)
+      |> Enum.map(fn filter ->
+        case filter do
+          [:is, "event:goal", {:event, name}] ->
+            [[:is, "event:name", name]]
+
+          [:is, "event:goal", {:page, page}] ->
+            [[:is, "event:page", page]]
+
+          [:member, "event:goal", events] ->
+            events
+            |> Enum.group_by(&elem(&1, 0), &elem(&1, 1))
+            |> Enum.map(fn
+              {:event, names} -> [:member, "event:name", names]
+              {:page, pages} -> [:member, "event:page", pages]
+            end)
+
+          filter ->
+            [filter]
+        end
+      end)
+      |> Enum.concat()
+
+    struct!(query, filters: new_filters)
+  end
+
+  defp do_decide_table(%Query{filters: [], property: nil}), do: "imported_visitors"
+  defp do_decide_table(%Query{filters: [], property: "event:props:url"}), do: nil
+  defp do_decide_table(%Query{filters: [], property: "event:props:path"}), do: nil
+
+  defp do_decide_table(%Query{filters: filters, property: "event:props:url"}) do
     case filters do
       [[:is, "event:name", name]] when name in @goals_with_url ->
         "imported_custom_events"
@@ -59,7 +130,7 @@ defmodule Plausible.Stats.Imported.Base do
     end
   end
 
-  def decide_table(%Query{filters: filters, property: "event:props:path"}) do
+  defp do_decide_table(%Query{filters: filters, property: "event:props:path"}) do
     case filters do
       [[:is, "event:name", name]] when name in @goals_with_path ->
         "imported_custom_events"
@@ -69,39 +140,30 @@ defmodule Plausible.Stats.Imported.Base do
     end
   end
 
-  def decide_table(%Query{filters: [], property: "event:goal"}) do
+  defp do_decide_table(%Query{filters: [], property: "event:goal"}) do
     "imported_custom_events"
   end
 
-  def decide_table(%Query{filters: [], property: property}) do
-    Imported.property_to_table_mappings()[property]
+  defp do_decide_table(%Query{filters: [], property: property}) do
+    @property_to_table_mappings[property]
   end
 
-  def decide_table(%Query{filters: filters, property: "event:goal"}) do
-    any_event_name_filters? =
-      filters
-      |> Enum.filter(fn filter -> Enum.at(filter, 1) in ["event:name"] end)
-      |> Enum.any?()
+  defp do_decide_table(%Query{filters: filters, property: "event:goal"}) do
+    filter_props = Enum.map(filters, &Enum.at(&1, 1))
 
-    any_page_filters? =
-      filters
-      |> Enum.filter(fn filter -> Enum.at(filter, 1) in ["event:page"] end)
-      |> Enum.any?()
-
-    any_other_filters? =
-      filters
-      |> Enum.filter(fn filter -> Enum.at(filter, 1) not in ["event:page", "event:name"] end)
-      |> Enum.any?()
+    any_event_name_filters? = "event:name" in filter_props
+    any_page_filters? = "event:page" in filter_props
+    any_other_filters? = Enum.any?(filter_props, &(&1 not in ["event:page", "event:name"]))
 
     cond do
       any_other_filters? -> nil
-      any_event_name_filters? && !any_page_filters? -> "imported_custom_events"
-      any_page_filters? && !any_event_name_filters? -> "imported_pages"
+      any_event_name_filters? and not any_page_filters? -> "imported_custom_events"
+      any_page_filters? and not any_event_name_filters? -> "imported_pages"
       true -> nil
     end
   end
 
-  def decide_table(%Query{filters: filters, property: property}) do
+  defp do_decide_table(%Query{filters: filters, property: property}) do
     table_candidates =
       filters
       |> Enum.map(fn [_, prop | _] -> prop end)
@@ -110,7 +172,7 @@ defmodule Plausible.Stats.Imported.Base do
         "visit:screen" -> "visit:device"
         prop -> prop
       end)
-      |> Enum.map(&Imported.property_to_table_mappings()[&1])
+      |> Enum.map(&@property_to_table_mappings[&1])
 
     case Enum.uniq(table_candidates) do
       [candidate] -> candidate
